+++ src/core_modules/capture-core/components/DataEntry/DataEntry.component.js
@@ -79,6 +79,7 @@ type FieldContainer = {
     field: React.Element<any>,
     placement: $Values<typeof placements>,
     section?: ?string,
+    sectionName?: ?string
 };
 
 type DirectionClasses = {
@@ -171,19 +172,17 @@ class DataEntryPlain extends React.Component<Props> {
             Object.keys(this.props.dataEntrySections).reduce((accSections, sectionKey) => {
                 const section = sections[sectionKey];
                 if (section.placement === placement) {
-                    const sectionFields = fields ?
-                        fields
-                            .filter(fieldContainer => fieldContainer.section === sectionKey)
-                            .map((fieldContainer, index, array) => (
-                                <FieldSection
-                                    formHorizontal={this.props.formHorizontal}
-                                    fieldContainer={fieldContainer}
-                                    index={index}
-                                    total={array.length}
-                                />
-                            ))
-                        : null;
-
+                    const sectionFields = fields
+                        .filter(fieldContainer => fieldContainer.section === sectionKey);
+                    const sectionFieldsContainer = sectionFields.map((fieldContainer, index, array) => (
+                        <FieldSection
+                            formHorizontal={this.props.formHorizontal}
+                            fieldContainer={fieldContainer}
+                            index={index}
+                            total={array.length}
+                        />
+                    ));
+                    const sectionFieldName = sectionFields.length && sectionFields[0].sectionName;
                     if (sectionFields && sectionFields.length > 0) {
                         accSections.push(
                             <div
@@ -193,11 +192,11 @@ class DataEntryPlain extends React.Component<Props> {
                                 <Section
                                     header={
                                         <SectionHeaderSimple
-                                            title={section.name}
+                                            title={sectionFieldName ?? section.name}
                                         />
                                     }
                                 >
-                                    {sectionFields}
+                                    {sectionFieldsContainer}
                                 </Section>
                             </div>,
                         );