+++ src/core_modules/capture-core/components/WidgetEnrollmentEventNew/Validated/useLifecycle.js
@@ -6,6 +6,7 @@ import type { OrgUnit } from '@dhis2/rules-engine-javascript';
 import { getOpenDataEntryActions, getRulesActions } from '../DataEntry';
 import type { TrackerProgram, ProgramStage, RenderFoundation } from '../../../metaData';
 import type { RulesExecutionDependenciesClientFormatted } from '../common.types';
+import { useCategoryCombinations } from '../../DataEntryDhis2Helpers/AOC/useCategoryCombinations';
 
 export const useLifecycle = ({
     program,
@@ -33,13 +34,17 @@ export const useLifecycle = ({
 
     const dataEntryReadyRef = useRef(false);
     const delayRulesExecutionRef = useRef(false);
+    const { programCategory, isLoading } = useCategoryCombinations(program.id);
+
     useEffect(() => {
-        dispatch(batchActions([
-            ...getOpenDataEntryActions(dataEntryId, itemId),
-        ]));
-        dataEntryReadyRef.current = true;
-        delayRulesExecutionRef.current = true;
-    }, [dispatch, dataEntryId, itemId, program, formFoundation]);
+        if (!isLoading) {
+            dispatch(batchActions([
+                ...getOpenDataEntryActions(dataEntryId, itemId, programCategory),
+            ]));
+            dataEntryReadyRef.current = true;
+            delayRulesExecutionRef.current = true;
+        }
+    }, [dispatch, dataEntryId, itemId, program, formFoundation, isLoading, programCategory]);
 
     const eventsRef = useRef();
     const attributesRef = useRef();
@@ -50,6 +55,7 @@ export const useLifecycle = ({
     // Refactor the helper methods (getCurrentClientValues, getCurrentClientMainData in rules/actionsCreator) to be more explicit with the arguments.
     const state = useSelector(stateArg => stateArg);
     useEffect(() => {
+        if (isLoading) { return; }
         if (delayRulesExecutionRef.current) {
             // getRulesActions depends on settings in the redux store that are being managed through getOpenDataEntryActions.
             // The purpose of the following lines of code is to make sure the redux store is ready before calling getRulesActions.
@@ -86,6 +92,7 @@ export const useLifecycle = ({
         dataEntryId,
         itemId,
         rulesExecutionTrigger,
+        isLoading,
     ]);
 
     const rulesReady =