+++ src/core_modules/capture-core/components/WidgetEventEdit/WidgetEventEdit.actions.js
@@ -7,5 +7,5 @@ export const actionTypes = {
         'WidgetEventEdit.StartShowEditEventDataEntry',
 };
 
-export const startShowEditEventDataEntry = (orgUnit: OrgUnit) =>
-    actionCreator(actionTypes.START_SHOW_EDIT_EVENT_DATA_ENTRY)({ orgUnit });
+export const startShowEditEventDataEntry = (orgUnit: OrgUnit, programCategory: Object) =>
+    actionCreator(actionTypes.START_SHOW_EDIT_EVENT_DATA_ENTRY)({ orgUnit, programCategory });