+++ src/reducers/descriptions/trackerCapture.reducerDescriptions.js
@@ -70,7 +70,7 @@ import { appReducerDesc } from './app.reducerDescription';
 import { currentSelectionsReducerDesc } from './currentSelections.reducerDescription';
 import { feedbackDesc } from './feedback.reducerDescription';
 import { activePageDesc } from '../../core_modules/capture-core/reducers/descriptions/activePage.reducerDescription';
-import { searchPageDesc } from '../../core_modules/capture-core/reducers/descriptions/searchPage.reducerDescription';
+import { searchDomainDesc } from '../../core_modules/capture-core/reducers/descriptions/searchDomain.reducerDescription';
 import { enrollmentPageDesc } from '../../core_modules/capture-core/reducers/descriptions/enrollmentPage.reducerDescription';
 import { enrollmentDomainDesc } from '../../core_modules/capture-core/reducers/descriptions/enrollmentDomain.reducerDescription';
 import { newPageDesc } from '../../core_modules/capture-core/reducers/descriptions/newPage.reducerDescription';
@@ -123,7 +123,7 @@ export const reducerDescriptions = [
     rulesEffectsGeneralErrorsDesc,
     rulesEffectsDisabledFieldsDesc,
     ruleEffectsOptionsVisibilityDesc,
-    searchPageDesc,
+    searchDomainDesc,
     teiSearchDesc,
     trackedEntityInstanceDesc,
     useNewDashboardDesc,